fix: logs and update UI to be more user friendly
